refactor: 💡 Fragmentのファイルにかかれているクラスを分離
