+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/CustomAdapter.kt
@@ -1,0 +1,40 @@
+package jp.co.yumemi.android.codecheck
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+
+/**
+ * RecyclerView用のAdapter
+ *
+ * @param itemClickListener OnItemClickListener Itemをクリックしたとき用Listener
+ */
+class CustomAdapter(
+    private val itemClickListener: OnItemClickListener,
+) : ListAdapter<Item, CustomAdapter.ViewHolder>(DiffUtil) {
+
+    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
+
+    interface OnItemClickListener {
+        fun itemClick(item: Item)
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val view = LayoutInflater.from(parent.context)
+            .inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        val item = getItem(position)
+        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text =
+            item.name
+
+        holder.itemView.setOnClickListener {
+            itemClickListener.itemClick(item)
+        }
+    }
+}
