+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/Item.kt
@@ -1,0 +1,15 @@
+package jp.co.yumemi.android.codecheck
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Item(
+    val name: String,
+    val ownerIconUrl: String,
+    val language: String,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val forksCount: Long,
+    val openIssuesCount: Long,
+) : Parcelable
