+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/SearchViewModel.kt
@@ -3,7 +3,6 @@
  */
 package jp.co.yumemi.android.codecheck
 
-import android.os.Parcelable
 import androidx.lifecycle.ViewModel
 import io.ktor.client.*
 import io.ktor.client.call.*
@@ -12,7 +11,6 @@ import io.ktor.client.statement.*
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
-import kotlinx.parcelize.Parcelize
 import org.json.JSONObject
 
 /**
@@ -76,14 +74,3 @@ class SearchViewModel(private val client: HttpClient) : ViewModel() {
         }.await()
     }
 }
-
-@Parcelize
-data class Item(
-    val name: String,
-    val ownerIconUrl: String,
-    val language: String,
-    val stargazersCount: Long,
-    val watchersCount: Long,
-    val forksCount: Long,
-    val openIssuesCount: Long,
-) : Parcelable
